Use Torsocks instead of passing socks url as ytdl argument for mpv
According to the documentation for mpv,

"A proxy URL can be passed for youtube-dl to use it in parsing the website.
This is useful for geo-restricted URLs. After youtube-dl parsing, some URLs
also require a proxy for playback, so this can pass that proxy information to
mpv. Take note that SOCKS proxies aren't supported and https URLs also bypass
the proxy. This is a limitation in FFmpeg."

Thus, the solution previously used only fetches some metadata with Tor,
but the actual content is then streamed over the clearnet.

Instead, I will now use Torsocks until I come up with a more elegant,
direct, solution.