Add query for maximum resolution, with lower default for Tor
Depending on how fast internet you have, you may want to use a lower
resolution than 1080, so 3 more options for maximum resolution are given
to the user. When using Tor, the default for this value is lower than
when not using Tor.

Also, the doSelectionQuery function has been improved by supporting
default values and by optionally returning either the index or the value
chosen.