Fix handling of consent page that breaks the application
The new version of the consent page that YouTube redirects new
users to breaks the application when trying to get HTML pages.

In this commit, I add an HTTP post to accept tracking (we throw
all cookies in the bin the next second anyway, so it shouldn't make
any difference).