Give correct diagnostics when YouTube rejects POST request
Sometimes, YouTube doesn't like the request we send for
accepting cookies, as it finds it malformed. This was
incorrectly diagnosed as there being no search results,
which is very misleading to the user. Until I can
figure out why it thinks the request is malformed, I will
give a more correct diagnostic instead.