Make control flow level function responsible for ReturnFromMenu
The function removeSubscriptionFromDatabaseByChannelTitle is not a
function intended to control application flow, and thus it should not
return ReturnFromMenu. This is because it is the function supposed to be
used in any context where a subscription should be removed from
database, not just when that is an option presented to the user in a
menu, which is the only context where ReturnFromMenu should be returned.

Instead, make an application control flow level function responsible for
calling removeSubscriptionFromDatabaseByChannelTitle and then returning
ReturnFromMenu.