change style when doing heavily nested list comprehension
When doing list comprehension where a large and possibly nested function
call is used as expression, put the for-clause on the same line as the
closing function call parenthesis. This makes it visually clearer that
the for-clause is part of the list comprehension, rather than the
beginning of a loop.