change style for heavily nested list initialization w.out comprehension
When doing a heavily nested list initialization, if there is a function
call with multiple parameters, the opening bracket should be on the same
line as the function identifier, all parameters should then be indented
one level further and take one line each. The closing bracket should
come on the first line after the last function parameter. If there are
more list elements/parameters after the one returned by the function
call, then the first of these shall be written on the same line as the
closing bracket of the function.