Refactor thumbnail fetching and http getting
It's stupid that the general method for getting HTTP content is called
something to do with HTML, and it's stupid that it automatically
converts gotten content to text. So I gave it a new name, and let it
return the response object instead. Then I used this function for
getting thumbnails instead of wget, so that tor circuits can be managed
internally again.