Make it certain that all threads are dead
By always starting threaded processes as separate processes, one can
kill the process to ensure that all the threads die (you cannot kill a
thread directly, for reasons that are probably good, but this is the
next best thing)

In general, embarking on the journey of threading has caused a lot of
instability, but the performance gains are definitely real. I'm going to
need to come up with a less ad hoc strategy for this though. I'm
thinking, maybe do a separate file where all multiprocessing stuff is
managed, so that it's easier to standardized. Right now it's fine
(kinda) because it's only a couple of places that I do threading, and
thus multiprocessing.