Fix multiprocessing-related bugs (mark as read, exceptions, constants)
This commit fixes a few issues related to the migration to
multiprocessing and threading. On some platforms, constants seem to not
be passed to new processes implicitly, so in this commit we make sure
they are passed explicitly, which should hopefully fix this problem.

I also made sure to raise exceptions if processes don't return
successfully.

I also noted that mark as read didn't work as expected; this is not
related to multiprocessing directly, but it is related to the migration;
it is because the database is no longer as persistent in memory, but
often read/written from/to file (to ensure thread safety). However, I
forgot to write the changes to file after marking entries as read, which
is fixed in this commit.