Begin refactoring into separate files instead of one huge monolith
The codebase is starting to become a bit big for just one file, which is
an issue when trying to get an overview of what is relevant for a
particular change. This commit takes care of some low hanging fruit that
can easily be broken out into separate files.

indicator_classes.py
parser_classes.py
presentation.py